Rename package from converter to migration.
