+++ google-cloud-tools-plugin/testSrc/com/google/cloud/tools/intellij/appengine/migration/AppEngineFacetMigrationConverterProviderTest.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2016 Google Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.cloud.tools.intellij.appengine.migration;
+
+import java.io.IOException;
+
+/**
+ * Unit tests for {@link AppEngineFacetMigrationConverterProvider}
+ */
+public class AppEngineFacetMigrationConverterProviderTest extends BaseConverterTest {
+
+  public void testConvertBasic() throws IOException {
+    testConvert("conversion/appEngineFacetMigration");
+  }
+
+}
