+++ google-cloud-tools-plugin/testSrc/com/google/cloud/tools/intellij/appengine/migration/BaseConverterTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.intellij.appengine.converter;
+package com.google.cloud.tools.intellij.appengine.migration;
 
 import com.intellij.appengine.AppEngineCodeInsightTestCase;
 import com.intellij.conversion.ProjectConversionTestUtil;
